moved sign out button to navbar, changed buttons to new color scheme
