+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import React, { Suspense, lazy } from 'react'
-import { Route, Switch } from 'react-router-dom'
+import { Route, Switch, useHistory } from 'react-router-dom'
 import { useLingui } from '@lingui/react'
 import { LandingPage } from './LandingPage'
 import { Main } from './Main'
@@ -8,7 +8,7 @@ import { TopBanner } from './TopBanner'
 import { PhaseBanner } from './PhaseBanner'
 import { Footer } from './Footer'
 import { Navigation } from './Navigation'
-import { Flex, Link, CSSReset } from '@chakra-ui/core'
+import { Flex, Link, CSSReset, useToast } from '@chakra-ui/core'
 import { SkipLink } from './SkipLink'
 import { TwoFactorNotificationBar } from './TwoFactorNotificationBar'
 import { useUserState } from './UserState'
@@ -32,7 +32,9 @@ const ResetPasswordPage = lazy(() => import('./ResetPasswordPage'))
 export default function App() {
   // Hooks to be used with this functional component
   const { i18n } = useLingui()
-  const { currentUser, isLoggedIn } = useUserState()
+  const toast = useToast()
+  const history = useHistory()
+  const { currentUser, isLoggedIn, logout } = useUserState()
 
   return (
     <>
@@ -51,30 +53,53 @@ export default function App() {
           <Link to="/">
             <Trans>Home</Trans>
           </Link>
-          <Link to="/domains">
-            <Trans>Domains</Trans>
-          </Link>
-
-          <Link to="/organizations">
-            <Trans>Organizations</Trans>
-          </Link>
-          {isLoggedIn() ? (
-            <Link to="/user">
-              <Trans>User Profile</Trans>
-            </Link>
-          ) : (
-            <Link to="/sign-in">
-              <Trans>Sign In</Trans>
-            </Link>
-          )}
 
           <Link to="/dmarc-report">
             <Trans>Report</Trans>
           </Link>
 
           {isLoggedIn() && (
+            <Link to="/domains">
+              <Trans>Domains</Trans>
+            </Link>
+          )}
+
+          {isLoggedIn() && (
+            <Link to="/organizations">
+              <Trans>Organizations</Trans>
+            </Link>
+          )}
+
+          {isLoggedIn() && (
+            <Link to="/user">
+              <Trans>User Profile</Trans>
+            </Link>
+          )}
+
+          {isLoggedIn() && (
             <Link to="/admin">
               <Trans>Admin Profile</Trans>
+            </Link>
+          )}
+          {isLoggedIn() ? (
+            <Link
+              onClick={() => {
+                logout()
+                history.push('/')
+                toast({
+                  title: 'Sign Out.',
+                  description: 'You have successfully been signed out.',
+                  status: 'success',
+                  duration: 9000,
+                  isClosable: true,
+                })
+              }}
+            >
+              Sign Out
+            </Link>
+          ) : (
+            <Link to="/sign-in">
+              <Trans>Sign In</Trans>
             </Link>
           )}
         </Navigation>
